feat: funcionalidad de donante cruzado

+++ backend/src/main/java/io/justina/justinaio/repositories/HistoriaClinicaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface HistoriaClinicaRepository extends JpaRepository<CasoClinico, Long>, JpaSpecificationExecutor<CasoClinico> {
+public interface HistoriaClinicaRepository extends JpaRepository<CasoClinico, Integer>, JpaSpecificationExecutor<CasoClinico> {
     Page<CasoClinico> findByPacienteIdPaciente(Integer idPaciente, Pageable pageable);
 }
